+++ src/tests/Resolution.test.ts
@@ -9,7 +9,7 @@ import Web3HttpProvider from 'web3-providers-http';
 import Web3WsProvider from 'web3-providers-ws';
 import Web3V027Provider from 'web3-0.20.7/lib/web3/httpprovider';
 import {
-  CryptoDomainWithAdaBchAddresses,
+  CryptoDomainWithAdaBchAddresses, CryptoDomainWithUsdtMultiChainRecords,
 } from '../utils/helpers';
 
 import {
@@ -338,6 +338,59 @@ describe('Resolution', () => {
           const lower = await resolution.addr('brad.crypto', 'eth');
           expectSpyToBeCalled(eyes, 2);
           expect(capital).toStrictEqual(lower);
+        });
+        describe('.multichain', () => {
+          it('should work with usdt on different erc20', async () => {
+            const erc20Spy = mockAsyncMethod(cns, "get", {
+              resolver: '0xb66DcE2DA6afAAa98F2013446dBCB0f4B0ab2842',
+              owner: '0xe7474D07fD2FA286e7e0aa23cd107F8379085037',
+              records: {
+                [standardKeys.USDT_ERC20]: '0xe7474D07fD2FA286e7e0aa23cd107F8379085037'
+              }
+            });
+            const erc20 = await resolution.multiChainAddr(CryptoDomainWithUsdtMultiChainRecords, "usdt", "erc20");
+            expect(erc20).toBe('0xe7474D07fD2FA286e7e0aa23cd107F8379085037');
+            expect(erc20Spy).toBeCalled();
+          });
+
+          it('should work with usdt tron chain', async () => {
+            const tronSpy = mockAsyncMethod(cns, "get", {
+              resolver: '0xb66DcE2DA6afAAa98F2013446dBCB0f4B0ab2842',
+              owner: '0xe7474D07fD2FA286e7e0aa23cd107F8379085037',
+              records: {
+                [standardKeys.USDT_TRON]: 'TNemhXhpX7MwzZJa3oXvfCjo5pEeXrfN2h'
+              }
+            });
+            const tron = await resolution.multiChainAddr(CryptoDomainWithUsdtMultiChainRecords, "usdt", "tron");
+            expect(tron).toBe('TNemhXhpX7MwzZJa3oXvfCjo5pEeXrfN2h');
+            expect(tronSpy).toBeCalled();
+          });
+
+          it('should work with usdt omni chain', async () => {
+            const omniSpy = mockAsyncMethod(cns, "get", {
+              resolver: '0xb66DcE2DA6afAAa98F2013446dBCB0f4B0ab2842',
+              owner: '0xe7474D07fD2FA286e7e0aa23cd107F8379085037',
+              records: {
+                [standardKeys.USDT_OMNI]: '19o6LvAdCPkjLi83VsjrCsmvQZUirT4KXJ'
+              }
+            });
+            const omni = await resolution.multiChainAddr(CryptoDomainWithUsdtMultiChainRecords, "usdt", "omni");
+            expect(omni).toBe('19o6LvAdCPkjLi83VsjrCsmvQZUirT4KXJ');
+            expect(omniSpy).toBeCalled();
+          });
+
+          it('should work with usdt eos chain', async () => {
+            const eosSpy = mockAsyncMethod(cns, "get", {
+              resolver: '0xb66DcE2DA6afAAa98F2013446dBCB0f4B0ab2842',
+              owner: '0xe7474D07fD2FA286e7e0aa23cd107F8379085037',
+              records: {
+                [standardKeys.USDT_EOS]: 'letsminesome'
+              }
+            });
+            const eos = await resolution.multiChainAddr(CryptoDomainWithUsdtMultiChainRecords, "usdt", "eos");
+            expect(eosSpy).toBeCalled();
+            expect(eos).toBe('letsminesome');
+          });
         });
       });