+++ CHANGELOG.md
@@ -9,6 +9,16 @@
   - remove type API
   - Introduced single config type SourceConfig
     
+## 1.20.0
+* Introduced Resolution#multiChainAddr(domain: string, ticker: string, chain: string) - More general method to fetch multi chain records.
+* Deprecated Resolution#usdt method in favor of multiChainAddr
+* Deprecated TickerVersion 
+
+## 1.19.0
+* Update @ensdomains/address-encoder dependency to remove security audit issues
+* Remove webpack dependency to remove security audit issues
+## 1.18.0
+* Use Infura Ethereum Provider by default
 ## 1.17.0
 * Add support for TLOS
 ## 1.16.2